requirements.txt
faiss-cpu
numpy
torch
PyPDF2
requests
pillow
tqdm
sentence-transformers 
flask